+++ echr-scraping/process_documents.py
@@ -17,8 +17,8 @@ from collections import Counter
 
 from gensim import corpora, models, similarities
 
-from NLP.data import load_text_file, load_CSV, data_transformations, match_city, department_name, max_n_gram, filter_per_inhabitants
-from NLP.preprocessing import rectify_missing_space, prepareText, frequencies, generateNGrams
+from nlp.NLP.data import load_text_file, load_CSV, data_transformations, match_city, department_name, max_n_gram, filter_per_inhabitants
+from nlp.NLP.preprocessing import rectify_missing_space, prepareText, frequencies, generateNGrams
 
 from nltk.tokenize import word_tokenize
 from nltk.util import ngrams
@@ -90,15 +90,16 @@ def main(args):
         print('Cannot load configuration file. Details: {}'.format(e))
         exit(5)
 
+    if not args.u:
+        try:
+            if args.f:
+                shutil.rmtree(args.output_folder)
+            os.mkdir(args.output_folder)
+        except Exception as e:
+            print(e)
+            exit(1)
 
-    try:
-        if args.f:
-            shutil.rmtree(args.output_folder)
-        os.mkdir(args.output_folder)
-    except Exception as e:
-        print(e)
-        exit(1)
-
+    update = args.u
     files = [os.path.join(args.input_folder, f) for f in listdir(args.input_folder) if isfile(join(args.input_folder, f)) if '_normalized.txt' in f]
     raw_corpus = []
     corpus_id = []
@@ -129,7 +130,9 @@ def main(args):
     #print(dictionary.token2id)
     corpus = [dictionary.doc2bow(text) for text in raw_corpus]
     for i, doc in enumerate(corpus):
-        with open(os.path.join(args.output_folder, '{}_bow.txt'.format(corpus_id[i])), 'w') as file:
+        filename = os.path.join(args.output_folder, '{}_bow.txt'.format(corpus_id[i]))
+        #if update and not os.path.isfile(filename):
+        with open(filename, 'w') as file:
             for f, v in doc:
                 file.write('{}:{} '.format(f, v))
 
@@ -154,6 +157,7 @@ if __name__ == "__main__":
     parser.add_argument('--output_folder', type=str, default="./processed_documents")
     parser.add_argument('--limit_tokens', type=int, default="100000")
     parser.add_argument('-f', action='store_true')
+    parser.add_argument('-u', action='store_true')
     args = parse_args(parser)
 
     main(args)
